apis updated

+++ backend/src/main/kotlin/com/burnoutstopper/burnouttest/repository/AnswerRepository.kt
@@ -1,11 +1,9 @@
 package com.burnoutstopper.burnouttest.repository
 
-import com.burnoutstopper.burnouttest.model.Answer
+import Answer
 import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.data.jpa.repository.Modifying
-import org.springframework.data.jpa.repository.Query
 
-interface AnswerRepository : JpaRepository<Answer, Long> {
+interface AnswerRepository : JpaRepository<Answer, Int> {
 
     fun findByRespondentId(respondentId: Int): Answer
 }
